bug fix, convert single to double

+++ src/RTHawk_gridding_recon_gpu.m
@@ -230,7 +230,7 @@ for s = 1:Nf
     cimg_nufft = complex(zeros(N1, N2, Nc, 'double'));
     for c = 1:Nc
         tstart = tic; fprintf('(%2d/%2d) NUFFT reconstruction (c=%2d/%2d)... ', s, Nf, c, Nc);
-        cimg_nufft(:,:,c) = nufft_adj(reshape(kspace(:,:,c,s) .* dcf, [Nk*Ni 1]), nufft_st) / sqrt(prod(Nd));
+        cimg_nufft(:,:,c) = nufft_adj(reshape(double(kspace(:,:,c,s)) .* double(dcf), [Nk*Ni 1]), double(nufft_st)) / sqrt(prod(Nd));
         fprintf('done! (%6.4f sec)\n', toc(tstart));
     end
